+++ packages/@tinacms/schema-tools/src/util/parseURL.spec.ts
@@ -6,11 +6,12 @@ import { parseURL, TINA_HOST } from './parseURL'
 const MOCK_HOST_PROD = `https://${TINA_HOST}`
 const MOCK_HOST_LOCAL = `https://localhost:4001`
 const MOCK_CLIENT_ID = '1234'
+const MOCK_VERSION = 'beta'
 const MOCK_BRANCH = 'main'
 
 describe('parseUrl', () => {
   it('returns the branch, client, ID and isLocalClient from a valid production URL', () => {
-    const correctURL = `${MOCK_HOST_PROD}/content/${MOCK_CLIENT_ID}/github/${MOCK_BRANCH}`
+    const correctURL = `${MOCK_HOST_PROD}/${MOCK_VERSION}/content/${MOCK_CLIENT_ID}/github/${MOCK_BRANCH}`
     const { branch, clientId, isLocalClient } = parseURL(correctURL)
     expect(branch).toBe(MOCK_BRANCH)
     expect(clientId).toBe(MOCK_CLIENT_ID)
@@ -25,7 +26,7 @@ describe('parseUrl', () => {
   })
   it('handles when branch name contains a "/"', () => {
     const MOCK_BRANCH = 'foo/bar'
-    const correctURL = `${MOCK_HOST_PROD}/content/${MOCK_CLIENT_ID}/github/${MOCK_BRANCH}`
+    const correctURL = `${MOCK_HOST_PROD}/${MOCK_VERSION}/content/${MOCK_CLIENT_ID}/github/${MOCK_BRANCH}`
     const { branch, clientId, isLocalClient } = parseURL(correctURL)
     expect(branch).toBe(MOCK_BRANCH)
     expect(clientId).toBe(MOCK_CLIENT_ID)