+++ packages/@tinacms/scripts/src/index.ts
@@ -472,6 +472,11 @@ export const buildIt = async (entryPoint, packageJSON) => {
 
   const outInfo = out(entry)
 
+  if (['@tinacms/app'].includes(packageJSON.name)) {
+    console.log('skipping @tinacms/app')
+    return
+  }
+
   external.forEach((ext) => (globals[ext] = 'NOOP'))
   if (target === 'node') {
     if (['@tinacms/graphql', '@tinacms/datalayer'].includes(packageJSON.name)) {
@@ -527,24 +532,24 @@ export const buildIt = async (entryPoint, packageJSON) => {
        * to do that with the latest version of @tinacms/mdx, it's easiest to inline the dependency
        * directly. Especially because this package bundles its dependencies
        */
-      const appMDXPath = path.join(
-        process.cwd(),
-        '..',
-        'app',
-        'appFiles',
-        'src',
-        'fields',
-        'rich-text',
-        'monaco',
-        'mdx.js'
-      )
-      await esbuild({
-        entryPoints: [path.join(process.cwd(), entry)],
-        bundle: true,
-        format: 'esm',
-        outfile: appMDXPath,
-        external: Object.keys({ ...peerDeps }),
-      })
+      // const appMDXPath = path.join(
+      //   process.cwd(),
+      //   '..',
+      //   'app',
+      //   'appFiles',
+      //   'src',
+      //   'fields',
+      //   'rich-text',
+      //   'monaco',
+      //   'mdx.js'
+      // )
+      // await esbuild({
+      //   entryPoints: [path.join(process.cwd(), entry)],
+      //   bundle: true,
+      //   format: 'esm',
+      //   outfile: appMDXPath,
+      //   external: Object.keys({ ...peerDeps }),
+      // })
     } else {
       await esbuild({
         entryPoints: [path.join(process.cwd(), entry)],